Add express to serve local http requests and avoid external dependencies on tests. Rename TestRecorder to HttpRecorder. Remove dependency between tests. Use async await in favor of promises for tests.
