+++ package.json
@@ -22,10 +22,12 @@
   "homepage": "https://github.com/comparaonline/nock-utils#readme",
   "devDependencies": {
     "@types/chai": "^4.0.4",
+    "@types/express": "^4.0.39",
     "@types/jsonfile": "^4.0.1",
     "@types/mocha": "^2.2.44",
     "@types/nock": "^8.2.1",
     "chai": "^4.1.2",
+    "express": "^4.16.2",
     "mocha": "^4.0.1",
     "nock": "^9.1.0",
     "ts-node": "^3.3.0",
